# docs/swagger.yaml
basePath: /api/tda/student-management
definitions:
...
  controller.ListProductResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/entity.Product'
        type: array
      message:
        type: string
      meta:
        $ref: '#/definitions/entity.Meta'
      success:
        type: boolean
    type: object
  controller.ProductDetailResponse:
    properties:
      data:
        $ref: '#/definitions/entity.Product'
      message:
        type: string
      success:
        type: boolean
    type: object
  controller.ProductResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
...
  controller.UpdateProduct:
    properties:
      name:
        type: string
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: This is list api for scheduling-api project
  license: {}
  title: Swagger scheduling-api project API
  version: "2.0"
paths:
...
  /product:
    post:
      consumes:
      - application/json
      description: Create a new product
      parameters:
      - description: product information
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/dto.ProductPayload'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ProductResponse'
      summary: Create a new product
      tags:
      - product-apis
  /product/{Id}:
    get:
      consumes:
      - application/json
      description: Get a product
      parameters:
      - description: product id
        in: path
        name: Id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ProductDetailResponse'
      summary: Get a product
      tags:
      - product-apis
  /product/active/{Id}:
    put:
      consumes:
      - application/json
      description: Active a product
      parameters:
      - description: product id
        in: path
        name: Id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ProductDetailResponse'
      summary: Active a product
      tags:
      - product-apis
  /product/deactive/{Id}:
    put:
      consumes:
      - application/json
      description: DeActive a product
      parameters:
      - description: product id
        in: path
        name: Id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ProductDetailResponse'
      summary: DeActive a product
      tags:
      - product-apis
  /product/update/{Id}:
    put:
      consumes:
      - application/json
      description: Update a product
      parameters:
      - description: product id
        in: path
        name: Id
        required: true
        type: integer
      - description: product information
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/controller.UpdateProduct'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ProductDetailResponse'
      summary: Update a product
      tags:
      - product-apis
  /products:
    get:
      consumes:
      - application/json
      description: Get list of products
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: pageSize
        in: query
        name: pageSize
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ListProductResponse'
      summary: Get list of products
      tags:
      - product-apis
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"